+++ web/src/home/home_presenter.dart
@@ -27,7 +27,7 @@ class HomePresenter implements mvp.Presenter {
   
   void addViewTo(ui.HasWidgets container) {
     assert(container is ui.TabLayoutPanel);
-    container.add(display.asWidget(), display.title);
+    (container as ui.TabLayoutPanel).addTabText(display.asWidget(), display.title);
   }
 }